refactor(client): share Solid server data (#42999)
