+++ packages/client/src/solid/index.ts
@@ -1,0 +1,2 @@
+export * from "./data"
+export * from "./connection"
